Fixing precicion problems, bluring shadow

+++ src/colors/color_mult.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   color_mult.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: klamprak <klamprak@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 11:28:34 by flfische          #+#    #+#             */
-/*   Updated: 2024/06/14 16:42:25 by flfische         ###   ########.fr       */
+/*   Updated: 2024/06/27 16:36:43 by klamprak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 #include <stdint.h>
 
-float	ft_color_clamp(float color)
+double	ft_color_clamp(double color)
 {
 	if (color < 0)
 		color = 0;
@@ -23,13 +23,13 @@ float	ft_color_clamp(float color)
 }
 
 /**
- * @brief Multiplies a color by a float value.
+ * @brief Multiplies a color by a double value.
  * @param color The color to multiply.
- * @param ratio The float value to multiply the color by.
+ * @param ratio The double value to multiply the color by.
  * @param result The color to store the result in.
  * @return t_color* The multiplied color.
  */
-t_color	*ft_color_float_mult(t_color color, float ratio, t_color *result)
+t_color	*ft_color_float_mult(t_color color, double ratio, t_color *result)
 {
 	result->r = color.r * ratio;
 	result->g = color.g * ratio;
