+++ src/maths/v3_dotprod.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   v3_dotprod.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: klamprak <klamprak@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/08 14:17:20 by flfische          #+#    #+#             */
-/*   Updated: 2024/06/13 10:28:14 by flfische         ###   ########.fr       */
+/*   Updated: 2024/06/27 16:36:43 by klamprak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@
  * @param b The second vector.
  * @return The dot product of the two vectors.
  */
-float	ft_v3_dotprod(const t_vector3 *a, const t_vector3 *b)
+double	ft_v3_dotprod(const t_vector3 *a, const t_vector3 *b)
 {
 	return (a->x * b->x + a->y * b->y + a->z * b->z);
 }
