+++ src/maths/v3_new.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   v3_new.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: flfische <flfische@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: klamprak <klamprak@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/08 14:14:31 by flfische          #+#    #+#             */
-/*   Updated: 2024/06/08 14:33:00 by flfische         ###   ########.fr       */
+/*   Updated: 2024/06/27 16:36:43 by klamprak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
  * @return The new vector. NULL if the allocation fails.
  * @note The function allocates memory.
  */
-t_vector3	*ft_v3_new(float x, float y, float z)
+t_vector3	*ft_v3_new(double x, double y, double z)
 {
 	t_vector3	*vector;
